rewrite 0.9 (beta) - now smaller & compacter scripts - more to separate functions - compacter modules - fix for eden module reassignto buildings after UNDO - modules now use config values - two interior scripts to one - split the templates out of the script - allow custom buildings templates from mission makers (ZEI_civ_customBuildings & ZEI_mil_customBuildings) - new module for easier creating a template/scheme without any code execute stuff - little optimizations - modules names are now more consistent - now the mod also supports buildings not only houses - the mod now allow also interiors for buildings without AI positions (experimental) - optimized the template generator (block CuratorCam) - some stuff i dont remember
