+++ ZEI/CfgFactionClasses.hpp
@@ -1,0 +1,6 @@
+class CfgFactionClasses {
+	class NO_CATEGORY;
+	class zei_interiors: NO_CATEGORY {
+		displayName = "Interiors (ZEI)";
+	};
+};
